Renames `TCSigner` to `TCSignerV3`
